apps/howtocook/pyproject.toml
[project]
name = "howtocook"
version = "0.1.0"
description = "HowToCook data scraper"
requires-python = ">=3.12"
dependencies = [
    "gh",
    "click>=8.0.0",
    "markdown-it-py>=3.0.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
]

[project.scripts]
howtocook = "howtocook.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/howtocook"]

[tool.uv.sources]
gh = { workspace = true }
